refactor: use screaming architecture & repositories + interface
