+++ app/auth/types/user.ts
@@ -1,0 +1,6 @@
+export interface UserServiceCreatePayload {
+  first_name: string
+  last_name: string
+  email: string
+  password: string
+}
